Add text search and clean up tweet middleware.

+++ migrations/20180913213014_initial_schema.js
@@ -24,7 +24,7 @@ exports.up = knex => knex.schema
       .notNullable();
 
     table
-      .bigInteger('retweetCount')
+      .integer('retweetCount')
       .notNullable();
 
     table
